--- livedata/livedata.go
@@ -1,2 +1,0 @@
-// Package livedata provides DDP reactive data source support.
-package livedata
