--- rpc/rpc.go
@@ -1,2 +1,0 @@
-// Package rpc provides DDP remote procedure call support.
-package rpc
